+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/records/RecordTypeConverter.kt
@@ -26,7 +26,7 @@ import kotlin.reflect.jvm.javaField
 class RecordTypeConverter<T : Record>(
   private val converterProvider: TypeConverterProvider,
   val type: KType
-) : DynamicAwareTypeConverters<T>(type.isMarkedNullable) {
+) : DynamicAwareTypeConverters<T>() {
   private val objectConstructorFactory = ObjectConstructorFactory()
   private val propertyDescriptors: Map<KProperty1<out Any, *>, PropertyDescriptor> by lazy {
     (type.classifier as KClass<*>)
@@ -49,12 +49,12 @@ class RecordTypeConverter<T : Record>(
   override fun convertFromDynamic(value: Dynamic, context: AppContext?, forceConversion: Boolean): T =
     exceptionDecorator({ cause -> RecordCastException(type, cause) }) {
       val jsMap = value.asMap()
-      return convertFromReadableMap(jsMap, context)
+      return convertFromReadableMap(jsMap, context, forceConversion)
     }
 
   override fun convertFromAny(value: Any, context: AppContext?, forceConversion: Boolean): T {
     if (value is ReadableMap) {
-      return convertFromReadableMap(value, context)
+      return convertFromReadableMap(value, context, forceConversion)
     }
 
     @Suppress("UNCHECKED_CAST")
@@ -65,7 +65,7 @@ class RecordTypeConverter<T : Record>(
 
   override fun isTrivial(): Boolean = false
 
-  private fun convertFromReadableMap(jsMap: ReadableMap, context: AppContext?): T {
+  private fun convertFromReadableMap(jsMap: ReadableMap, context: AppContext?, forceConversion: Boolean): T {
     val kClass = type.classifier as KClass<*>
     val instance = getObjectConstructor(kClass).construct()
 
@@ -85,7 +85,7 @@ class RecordTypeConverter<T : Record>(
           val javaField = property.javaField!!
 
           val casted = exceptionDecorator({ cause -> FieldCastException(property.name, property.returnType, type, cause) }) {
-            descriptor.typeConverter.convert(this, context)
+            descriptor.typeConverter.convert(this, context, forceConversion)
           }
 
           if (casted != null) {