+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/types/AnyType.kt
@@ -317,8 +317,9 @@ class AnyType(
     }
   }
 
-  fun convert(value: Any?, appContext: AppContext? = null): Any? =
-    converter.convert(value, appContext)
+  fun convert(value: Any?, appContext: AppContext? = null, forceConversion: Boolean = false): Any? {
+    return converter.convert(value, appContext, forceConversion)
+  }
 
   fun getCppRequiredTypes(): ExpectedType = converter.getCppRequiredTypes()