+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/types/ColorTypeConverter.kt
@@ -171,9 +171,7 @@ private val namedColors = mapOf(
 }
 
 @RequiresApi(Build.VERSION_CODES.O)
-class ColorTypeConverter(
-  isOptional: Boolean
-) : DynamicAwareTypeConverters<Color>(isOptional) {
+class ColorTypeConverter : DynamicAwareTypeConverters<Color>() {
   override fun convertFromDynamic(value: Dynamic, context: AppContext?, forceConversion: Boolean): Color {
     return when (value.type) {
       ReadableType.Number -> colorFromInt(value.asDouble().toInt())