+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/types/DurationTypeConverter.kt
@@ -9,7 +9,7 @@ import kotlin.time.Duration
 import kotlin.time.DurationUnit
 import kotlin.time.toDuration
 
-class DurationTypeConverter(isOptional: Boolean) : DynamicAwareTypeConverters<Duration>(isOptional) {
+class DurationTypeConverter : DynamicAwareTypeConverters<Duration>() {
   override fun convertFromDynamic(value: Dynamic, context: AppContext?, forceConversion: Boolean): Duration {
     if (value.type != ReadableType.Number) {
       throw IllegalArgumentException("Expected a number, but received ${value.type}")