+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/types/TypeConverter.kt
@@ -11,11 +11,11 @@ import expo.modules.kotlin.jni.ExpectedType
  * Basic type converter. It has to handle two different inputs - [Dynamic] and [Any].
  * The first one is used in the bridge implementation. The second one is used in the JSI.
  */
-abstract class TypeConverter<Type : Any> {
+interface TypeConverter<Type : Any> {
   /**
    * Tries to convert from [Any]? (can be also [Dynamic]) to the desired type.
    */
-  abstract fun convert(value: Any?, context: AppContext? = null, forceConversion: Boolean = false): Type?
+  fun convert(value: Any?, context: AppContext? = null, forceConversion: Boolean = false): Type?
 
   /**
    * Returns a list of [ExpectedType] types that can be converted to the desired type.
@@ -23,37 +23,22 @@ abstract class TypeConverter<Type : Any> {
    * For instance js object can be pass as [Map] or [expo.modules.kotlin.jni.JavaScriptObject].
    * This value tells us which one we should choose.
    */
-  open fun getCppRequiredTypes(): ExpectedType = ExpectedType(CppType.ANY)
+  fun getCppRequiredTypes(): ExpectedType = ExpectedType(CppType.ANY)
 
   /**
    * Checks if the current converter is a trivial one.
    * In that context, a trivial converter indicates a converter that in JSI implementation does nothing.
    * It should be true for most classes.
    */
-  open fun isTrivial(): Boolean = true
+  fun isTrivial(): Boolean = true
 }
 
-abstract class NullAwareTypeConverter<Type : Any>(
-  /**
-   * Whether `null` can be assigned to the desired type.
-   */
-  private val isOptional: Boolean
-) : TypeConverter<Type>() {
-  override fun convert(value: Any?, context: AppContext?, forceConversion: Boolean): Type? {
-    if (value == null || value is Dynamic && value.isNull) {
-      if (isOptional) {
-        return null
-      }
-      throw NullArgumentException()
-    }
-    return convertNonOptional(value, context, forceConversion)
+abstract class NonNullableTypeConverter<Type : Any>() : TypeConverter<Type> {
+  override fun convert(value: Any?, context: AppContext?, forceConversion: Boolean): Type {
+    return convertNonNullable(value ?: throw NullArgumentException(), context, forceConversion)
   }
 
-  /**
-   * Tries to convert from [Any] to the desired type.
-   * We know in that place that we're not dealing with `null`.
-   */
-  abstract fun convertNonOptional(value: Any, context: AppContext?, forceConversion: Boolean): Type
+  abstract fun convertNonNullable(value: Any, context: AppContext?, forceConversion: Boolean): Type
 }
 
 /**
@@ -61,8 +46,8 @@ abstract class NullAwareTypeConverter<Type : Any>(
  * Right it is used as a default base class for all converters, but this will change when we
  * stop using the bridge to pass data between JS and Kotlin.
  */
-abstract class DynamicAwareTypeConverters<T : Any>(isOptional: Boolean) : NullAwareTypeConverter<T>(isOptional) {
-  override fun convertNonOptional(value: Any, context: AppContext?, forceConversion: Boolean): T =
+abstract class DynamicAwareTypeConverters<T : Any>() : NonNullableTypeConverter<T>() {
+  override fun convertNonNullable(value: Any, context: AppContext?, forceConversion: Boolean): T =
     if (value is Dynamic) {
       convertFromDynamic(value, context, forceConversion)
     } else {
@@ -74,11 +59,10 @@ abstract class DynamicAwareTypeConverters<T : Any>(isOptional: Boolean) : NullAw
 }
 
 inline fun <reified T : Any> createTrivialTypeConverter(
-  isOptional: Boolean,
   cppRequireType: ExpectedType,
   crossinline dynamicFallback: (Dynamic) -> T = { throw UnsupportedClass(T::class) }
 ): TypeConverter<T> {
-  return object : DynamicAwareTypeConverters<T>(isOptional) {
+  return object : DynamicAwareTypeConverters<T>() {
     override fun convertFromDynamic(value: Dynamic, context: AppContext?, forceConversion: Boolean): T {
       return dynamicFallback(value)
     }